+++ ast/ast.cpp
@@ -26,6 +26,7 @@ void LiteralExpr::accept(ASTVisitor* visitor) { visitor->visit(this); }
 void IdentExpr::accept(ASTVisitor* visitor) { visitor->visit(this); }
 void AllocExpr::accept(ASTVisitor* visitor) { visitor->visit(this); }
 void InterfaceDecl::accept(ASTVisitor* visitor) { visitor->visit(this); }
+void ContinueStmt::accept(ASTVisitor* visitor) { visitor->visit(this); }
 void SwitchStmt::accept(ASTVisitor* visitor) { visitor->visit(this); }
 void TemplateCallExpr::accept(ASTVisitor* visitor) { visitor->visit(this); }
 void StructInitExpr::accept(ASTVisitor* visitor) { visitor->visit(this); }